Sparse vector - Int to Long

+++ pinecone-core/src/main/scala/io/cequence/pineconescala/domain/SparseVector.scala
@@ -1,6 +1,6 @@
 package io.cequence.pineconescala.domain
 
 case class SparseVector(
-  indices: Seq[Int],
+  indices: Seq[Long],
   values: Seq[Double]
 )
